backend/requirements.txt
fastapi
uvicorn
openai-whisper
torch
transformers
python-multipart
ffmpeg-python

pytest
httpx
